Detailed commit entry
ServerSide

* Added common.lang
* Analytics running framework. This was a pia due to hibernate bytecode manipulation and threading
* Added view framework for responses
* Sessions and attendance DB components
* Module endpoints

ClientSide

* Started work on user profile. Displays module info for students
* Module Admin screens
	- Search screen
	- Overview Screen
* Module Student screen, now show assignment information
* New loading animation on panels while loading
* Added plotlyJs for graphing

+++ Server/src/main/java/com/tom_maxwell/project/modules/assignments/AssignmentMarkModel.java
@@ -25,7 +25,7 @@ public class AssignmentMarkModel {
 	@JsonManagedReference
 	private AssignmentModel assignment;
 
-	private int percentage;
+	private double percentage;
 
 	public AssignmentMarkModel() {
 	}
@@ -60,11 +60,11 @@ public class AssignmentMarkModel {
 		this.assignment = assignment;
 	}
 
-	public int getPercentage() {
+	public double getPercentage() {
 		return percentage;
 	}
 
-	public void setPercentage(int percentage) {
+	public void setPercentage(double percentage) {
 		this.percentage = percentage;
 	}
 }
